pyproject.toml
[tool.setuptools.packages.find]
where = ["."]
exclude = ["data*", "history*", "build*", "*.egg-info*"]

[project]
name = "sandy"
version = "0.2.0"
description = "Local Discord personality AI with memory, tools, and observability"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.6.0",
    "discord.py[voice]>=2.7.1",
    "httpx>=0.28.1",
    "nvidia-ml-py>=12.570.86",
    "ollama>=0.4.0",
    "pillow>=11.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.0.0",
    # discord 2.7.1 broke voice-recv extension and this PR fixes it but hasn't been merged yet.  Fix later after merge.
    "discord-ext-voice-recv @ git+https://github.com/rdphillips7/discord-ext-voice-recv.git@ddd28601fe556f585b869e215f29c8236b95f88f",
    "faster-whisper",
    "nvidia-cublas-cu12",
    "nvidia-cudnn-cu12==9.*",
    "PyNaCl>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore:'audioop' is deprecated and slated for removal in Python 3.13:DeprecationWarning:discord\\.player",
]
